+++ zia/resource_zia_firewall_ips_rules.go
@@ -136,8 +136,7 @@ func resourceFirewallIPSRules() *schema.Resource {
 				Type:     schema.TypeSet,
 				Optional: true,
 				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validateDestAddress, // Apply the custom validation function here
+					Type: schema.TypeString,
 				},
 				Description: "Destination addresses. Supports IPv4, FQDNs, or wildcard FQDNs",
 			},