+++ zia/resource_zia_nat_control_rules.go
@@ -129,8 +129,7 @@ func resourceNatControlRules() *schema.Resource {
 				Type:     schema.TypeSet,
 				Optional: true,
 				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validateDestAddress,
+					Type: schema.TypeString,
 				},
 				Description: "Destination addresses. Supports IPv4, FQDNs, or wildcard FQDNs",
 			},